this demo shows how to numerically solve the clos
this demo numerically solves for the closest solution by mapping the input to the library of points. The first version is a two by two grid with nine points pre-calculated and stored. By calculating the new point and comparing it with each of the point in the library, the closest solution is acheived at the point where they have smallest difference. This version has not implemented ways to differentiate two points with the same difference from the sound source point. When storing more data points, the system will be more accuracy.